feat: update GPT-3 integration to gpt-3.5-turbo

+++ services/index.js
@@ -2,7 +2,7 @@ import config from '../config'
 import { replyMessage } from '../utils/replyMessage'
 import { aiAssistant } from './aiAssistant'
 
-const { DISCORD_CHANNEL_ID, DISCORD_CHANNEL_MAX_MESSAGE } = config
+const { DISCORD_CHANNEL_ID, DISCORD_CHANNEL_MAX_MESSAGE, OPEN_AI_GPT_PERSONA } = config
 const { defaultReplyMessage, errorMessage } = replyMessage()
 
 /**
@@ -12,27 +12,52 @@ const { defaultReplyMessage, errorMessage } = replyMessage()
 export async function messageHandler(message) {
   console.log(`「${message.channel.name}」${message.author.username}：${message.content} `)
 
+  // 只有在自己發話時會產生回應 for development
+  if (message.author.tag !== 'Sheep#2929') return
+
   if (message.channel.id === DISCORD_CHANNEL_ID) {
     let cacheMsg = null
     try {
       cacheMsg = await message.channel.send(defaultReplyMessage)
 
-      let prompt = ''
+      let messages = []
       if (DISCORD_CHANNEL_MAX_MESSAGE === 1) {
-        prompt = message.content
+        messages.push({
+          role: 'user',
+          content: message.content,
+        })
       } else {
         const channelMessageData = await message.channel.messages.fetch({
           limit: DISCORD_CHANNEL_MAX_MESSAGE + 1,
         })
-        const channelMessageHistory = channelMessageData.map(msg => msg.content)
-        prompt = channelMessageHistory
-          .reverse()
-          .filter(msg => msg !== defaultReplyMessage && msg !== '')
-          .join('\n')
+        const reverseMessages = channelMessageData.reverse()
+        messages = reverseMessages.map((msg) => {
+          if (msg.author.bot) {
+            if (msg.content === defaultReplyMessage)
+              return null
+
+            return {
+              role: 'system',
+              content: msg.content,
+            }
+          }
+
+          return {
+            role: 'user',
+            content: msg.content,
+          }
+        }).filter(msg => msg !== null)
       }
 
-      const aiResponse = await aiAssistant(prompt)
+      messages.unshift({
+        role: 'user',
+        content: OPEN_AI_GPT_PERSONA,
+      })
+
+      const aiResponse = await aiAssistant(messages)
+
       await cacheMsg.delete()
+
       message.reply(aiResponse)
     } catch (error) {
       console.log(error)
